Make sensors useful in cardGrid
Previously: sensors in cardGrid show an icon and the title below.

Desired Result: Sensors now display the state in place of the icon.

Rationale: Displaying a static icon in cardGrid is pointless. This way the state is displayed in large font which makes it useful and easier to read for the sight impaired.

Synopsis of Changes: render_card() now passes on cardType to generate_entities_page() which in turn passes it onto generate_entities_item(). If the cardType is `cardGrid` then the state is displayed instead of the icon.